+++ TP2/Userland/SampleCodeModule/include/Speaker.h
@@ -1,9 +1,10 @@
 #ifndef SPEAKER_H
 #define SPEAKER_H
+#include <stdint.h>
 
 void stopSound();
 
-void playSound(int frecuency);
+void playSound(uint64_t frecuency);
 
 
 #endif
